shadow mapping is basically working, but there are still some quirks and  it's not optimized at all (very poor performance)

+++ neo/renderer/Material.h
@@ -288,7 +288,7 @@ struct glslProgramDef_t {
   char   vertexShaderName[64];
   char   fragmentShaderName[64];
   GLuint ident;
-  byte   usedTextureUnits;
+  unsigned short  usedTextureUnits;
 };
 
 extern const glslProgramDef_t* shadowProgram;
